+++ src/main/java/com/ticketing/ticketing_system/TicketingSystemApplication.java
@@ -2,8 +2,10 @@ package com.ticketing.ticketing_system;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 
 @SpringBootApplication
+@EnableCaching
 public class TicketingSystemApplication {
 
 	public static void main(String[] args) {
