+++ src/main/java/com/ticketing/ticketing_system/controllers/TicketController.java
@@ -4,6 +4,9 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,9 +34,11 @@ public class TicketController {
     }
 
     // Get all tickets
+    @Cacheable("tickets")
     @GetMapping("/tickets")
-    public List<Ticket> fetchAllTickets() {
-        return ticketRepository.findAll();
+    public Page<Ticket> fetchAllTickets(@RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "5") int size) {
+        PageRequest pageable = PageRequest.of(page,size);
+        return ticketRepository.findAll(pageable);
     }
 
     // Get a specific ticket
